Arreglada doIntersect

+++ World/Calculus/Calculus.cpp
@@ -16,7 +16,7 @@ int orientation(dpoint_t p, dpoint_t q, dpoint_t r)
 {
 	// See http://www.geeksforgeeks.org/orientation-3-ordered-points/
 	// for details of below formula.
-	int val = (int)((q.y - p.y) * (r.x - q.x) -
+	double val = (double)((q.y - p.y) * (r.x - q.x) -
 		(q.x - p.x) * (r.y - q.y));
 
 	if (val == 0)
@@ -29,10 +29,10 @@ bool doIntersect(dpoint_t p1, dpoint_t q1, dpoint_t p2, dpoint_t q2)	//Funcion q
 {
 	// Find the four orientations needed for general and
 	// special cases
-	int o1 = orientation(p1, q1, p2);
-	int o2 = orientation(p1, q1, q2);
-	int o3 = orientation(p2, q2, p1);
-	int o4 = orientation(p2, q2, q1);
+	double o1 = orientation(p1, q1, p2);
+	double o2 = orientation(p1, q1, q2);
+	double o3 = orientation(p2, q2, p1);
+	double o4 = orientation(p2, q2, q1);
 
 	// General case
 	if (o1 != o2 && o3 != o4)
@@ -53,6 +53,88 @@ bool doIntersect(dpoint_t p1, dpoint_t q1, dpoint_t p2, dpoint_t q2)	//Funcion q
 
 	return false; // Doesn't fall in any of the above cases
 }
+
+//bool doIntersect(dpoint_t p0, dpoint_t p1, dpoint_t p2, dpoint_t p3)
+//{
+//	double x0, y0, x1, y1, x2, y2, x3, y3;
+//	double m1, m2;
+//	double x, y;
+//	bool answer;
+//	x0 = p0.x;
+//	y0 = p0.y;
+//	x1 = p1.x;
+//	y1 = p1.y;
+//	x2 = p2.x;
+//	y2 = p2.y;
+//	x3 = p3.x;
+//	y3 = p3.y;
+//
+//	if (x0 != x1)
+//		m1 = calculatePendient(p0, p1);
+//	if (x2 != x3)
+//		m2 = calculatePendient(p2, p3);
+//	if (x0 != x1 && x2 != x3)
+//	{
+//		if (m1 != m2)
+//		{
+//			x = (y2 - m2*x2 - y0 + m1*x0) / (m1 - m2);
+//			y = m1*x + y0 - m1*x0;
+//			answer = (isItInside(x2, y2, x3, y3, x, y)) && (isItInside(x0, y0, x1, y1, x, y));//ME FIJO QUE EL X Y EL Y OBTENIDOS ESTEN DENTRO DE LOS SEGMENTOS
+//		}
+//		else //SI LAS PENDIENTES SON LAS MISMAS
+//			answer = false;
+//	}
+//	if (x0 == x1 && x2 != x3)
+//	{
+//		x = x0;
+//		y = m2 * x + y2 - m2*x2;
+//		answer = (isItInside(x2, y2, x3, y3, x, y));
+//	}
+//	else if (x0 != x1 && x2 == x3)
+//	{
+//		x = x2;
+//		y = m1 * x + y0 - m1*x0;
+//		answer = (isItInside(x0, y0, x1, y1, x, y));
+//	}
+//	else
+//		answer = false;
+//	return answer;
+//}
+//
+//bool isItInside(double x0, double y0, double x1, double y1, double x, double y)
+//{
+//	if (x1 >= x0 && y1 >= y0)
+//	{
+//		if (x <= x1 && x >= x0 && y <= y1 && y >= y0)
+//			return true;
+//		else
+//			return false;
+//	}
+//	else if (x1 <= x0 && y1 >= y0)
+//	{
+//		if (x <= x0 && x >= x1 && y <= y1 && y >= y0)
+//			return true;
+//		else
+//			return false;
+//	}
+//	else if (x1 >= x0 && y1 <= y0)
+//	{
+//		if (x <= x1 && x >= x0 && y <= y0 && y >= y1)
+//			return true;
+//		else
+//			return false;
+//	}
+//	else if (x1 <= x0 && y1 <= y0)
+//	{
+//		if (x <= x0 && x >= x1 && y <= y0 && y >= y1)
+//			return true;
+//		else
+//			return false;
+//	}
+//	else
+//		return false;
+//}
+
 
 double max(double x, double y)
 {
